Add logs to identify the issue
Update HLs dlete logic

Remove unused changes

+++ android/java/org/chromium/chrome/browser/playlist/hls_content/HlsServiceImpl.java
@@ -84,7 +84,11 @@ public class HlsServiceImpl extends HlsService.Impl implements ConnectionErrorHa
                             playlistItemId,
                             playlistItem -> {
                                 if (playlistItem == null) {
-                                    removeContentAndStartNextDownload(playlistItemId);
+                                    PostTask.postTask(
+                                            TaskTraits.USER_VISIBLE_MAY_BLOCK,
+                                            () -> {
+                                                removeContentAndStartNextDownload(playlistItemId);
+                                            });
                                 }
                                 currentDownloadingPlaylistItemId = playlistItemId;
                                 HlsUtils.getManifestFile(
@@ -162,6 +166,9 @@ public class HlsServiceImpl extends HlsService.Impl implements ConnectionErrorHa
 
     private void removeContentAndStartNextDownload(String playlistItemId) {
         PlaylistRepository playlistRepository = new PlaylistRepository(mContext);
+        if (playlistRepository == null) {
+            return;
+        }
         playlistRepository.deleteHlsContentQueueModel(playlistItemId);
         currentDownloadingPlaylistItemId = "";
         if (playlistRepository.getFirstHlsContentQueueModel() != null) {
@@ -191,7 +198,9 @@ public class HlsServiceImpl extends HlsService.Impl implements ConnectionErrorHa
                                     playlistItem.mediaFileBytes,
                                     playlistItem.cached,
                                     false);
-                    VideoPlaybackService.Companion.addNewPlaylistItemModel(playlistItemModel);
+                    if (HlsUtils.isVideoPlaybackServiceRunning()) {
+                        VideoPlaybackService.Companion.addNewPlaylistItemModel(playlistItemModel);
+                    }
                 });
     }
 
@@ -211,5 +220,13 @@ public class HlsServiceImpl extends HlsService.Impl implements ConnectionErrorHa
 
         mPlaylistService =
                 PlaylistServiceFactoryAndroid.getInstance().getPlaylistService(HlsServiceImpl.this);
+    }
+
+    @Override
+    public void onDestroy() {
+        if (mPlaylistService != null) {
+            mPlaylistService.close();
+        }
+        super.onDestroy();
     }
 }
